aws-api/requirements.txt
flask-restful
boto3